+++ src/components/Balance/Balance.styled.js
@@ -1,0 +1,14 @@
+import styled from "styled-components";
+
+const StyledBalance = styled.div`
+  width: 100%;
+  background-color: #c0eda6;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  height: 70px;
+  border: none;
+  margin-top: 10px;
+`;
+
+export { StyledBalance };
