+++ src/components/ToggleDisplay/ToggleDisplay.jsx
@@ -1,16 +1,6 @@
-import styled from "styled-components";
 import { Button } from "components";
 import { useDisplay } from "hooks";
-
-const StyledDisplay = styled.div`
-  position: absolute;
-  top: 30px;
-  right: 100px;
-
-  button {
-    box-shadow: 0px 0px 7px;
-  }
-`;
+import { StyledDisplay } from "./ToggleDisplay.styled";
 
 function ToggleDisplay() {
   const { displayMode, toggleDisplay } = useDisplay();
