+++ src/components/ToggleDisplay/ToggleDisplay.styled.js
@@ -1,0 +1,13 @@
+import styled from "styled-components";
+
+const StyledDisplay = styled.div`
+  position: absolute;
+  top: 30px;
+  right: 100px;
+
+  button {
+    box-shadow: 0px 0px 7px;
+  }
+`;
+
+export { StyledDisplay };
