+++ src/components/vend-controller/HistoryBox.jsx
@@ -1,17 +1,6 @@
 import { HistoryContext } from "context";
 import { useContext } from "react";
-import styled from "styled-components";
-
-const StyledHistoryBox = styled.ul`
-  height: 300px;
-  background-color: #f2f2f2;
-  overflow-y: scroll;
-  margin-top: 10px;
-
-  li {
-    padding: 5px;
-  }
-`;
+import { StyledHistoryBox } from "./VendController.styled";
 
 function History({ comment }) {
   return comment;
