+++ src/components/vend-controller/InsertCoin.jsx
@@ -1,23 +1,6 @@
-import styled from "styled-components";
 import { useContext } from "react";
-import { CorrectCoinContext, InsertCoinContext, SetInsertCoinContext, AddHistoryContext } from "context";
-import { useEffect } from "react";
-import { useTimer } from "hooks";
-
-const StyledInsertCoin = styled.input`
-  width: 100%;
-  height: 70px;
-  background-color: #f2f2f2;
-  border: none;
-  border-radius: 10px;
-  margin: 0px 10px 10px 0px;
-  text-align: center;
-  font-size: 1.5em;
-
-  &:focus {
-    outline: none;
-  }
-`;
+import { AddHistoryContext, CorrectCoinContext, InsertCoinContext, SetInsertCoinContext } from "context";
+import { StyledInsertCoin, TotalInsertCoin } from "./VendController.styled";
 
 const TotalInsertCoin = styled.p`
   color: #f2f2f2;
@@ -29,9 +12,8 @@ const TotalInsertCoin = styled.p`
 function InsertCoin() {
   const correctCoin = useContext(CorrectCoinContext);
   const insertCoin = useContext(InsertCoinContext);
-  const setInsertCoin = useContext(SetInsertCoinContext);
   const addHistory = useContext(AddHistoryContext);
-  const { setDebounce } = useTimer();
+  const setInsertCoin = useContext(SetInsertCoinContext);
 
   const handleInsertCoinBlur = ({ target }) => {
     const inputCoin = Number(target.value);
@@ -44,18 +26,6 @@ function InsertCoin() {
       coin: correctedCoin,
     });
   };
-
-  useEffect(() => {
-    if (!insertCoin) return;
-    const delaySelectTime = 5000;
-
-    const autoReturn = () => {
-      setInsertCoin(0);
-      addHistory("RETURN_COIN", { change: insertCoin });
-    };
-
-    setDebounce(autoReturn, delaySelectTime);
-  }, [setInsertCoin, addHistory, insertCoin]);
 
   return (
     <>
