+++ src/context/CoinProvider.jsx
@@ -1,5 +1,6 @@
-import React, { createContext } from "react";
-import { useCoin } from "hooks";
+import React, { createContext, useEffect, useContext } from "react";
+import { useCoin, useTimer } from "hooks";
+import { SetInsertCoinContext, InsertCoinContext, AddHistoryContext } from "context";
 import { money } from "data";
 
 const CoinContext = createContext();
@@ -8,6 +9,21 @@ const CorrectCoinContext = createContext();
 
 function CoinProvider({ children }) {
   const { coin, selectCoin, correctCoin } = useCoin(money);
+  const { setDebounce } = useTimer();
+  const insertCoin = useContext(InsertCoinContext);
+  const addHistory = useContext(AddHistoryContext);
+  const setInsertCoin = useContext(SetInsertCoinContext);
+
+  const autoReturn = () => {
+    setInsertCoin(0);
+    addHistory("RETURN_COIN", { change: insertCoin });
+  };
+
+  useEffect(() => {
+    if (!insertCoin) return;
+    const delaySelectTime = 3000;
+    setDebounce(autoReturn, delaySelectTime);
+  }, [insertCoin]);
 
   return (
     <CoinContext.Provider value={coin}>
