+++ src/hooks/useCoin.jsx
@@ -1,56 +1,59 @@
 import { useState, useCallback } from "react";
+import { near } from "utils";
 
 function useCoin(init) {
   const [coin, setCoin] = useState(init);
 
+  const decreaseCount = (prevCoin, unit) =>
+    prevCoin.map((current) => {
+      if (current.unit === unit) {
+        return { ...current, count: current.count - 1 };
+      }
+      return current;
+    });
+
   const selectCoin = useCallback((unit) => {
+    setCoin((prevCoin) => decreaseCount(prevCoin, unit));
+  }, []);
+
+  const initCoin = (totalCoin) => {
     setCoin((prevCoin) =>
       prevCoin.map((current) => {
-        if (current.unit === unit) {
-          return { ...current, count: current.count - 1 };
-        }
-        return current;
+        return { ...current, count: 0 };
       })
     );
-  }, []);
+    return totalCoin;
+  };
 
   const correctCoin = useCallback(
     (inputCoin) => {
+      let copyCoin = coin;
+      const totalCoin = coin.reduce((acc, { unit, count }) => acc + unit * count, 0);
       let acc = 0;
 
-      const coinArr = coin.map((obj) => {
-        return { ...obj, total: obj.unit * obj.count };
-      });
-
-      const totalCoin = coinArr.reduce((acc, { total }) => acc + total, 0);
-      if (inputCoin > totalCoin) {
-        setCoin((prevCoin) =>
-          prevCoin.map((current) => {
-            return { ...current, count: 0 };
-          })
-        );
-        return totalCoin;
-      }
+      if (inputCoin > totalCoin) return initCoin(totalCoin);
 
       function recursive(remainCoin) {
-        for (let i = 0; i < coinArr.length; i++) {
-          const { unit, count, total } = coinArr[i];
+        const unitArr = copyCoin.map(({ count, unit }) => count && unit).filter((e) => e !== 0);
+        const unit = near(unitArr, remainCoin);
 
-          if (inputCoin <= acc) return acc;
-          else if (Math.floor(total / remainCoin) > 0) {
-            if (count) {
-              selectCoin(unit);
-              acc += unit;
-              const result = recursive(remainCoin - unit);
-              return result;
-            }
-          }
+        const nearCoin = copyCoin.find((e) => e.unit === unit).unit;
+
+        acc += nearCoin;
+        copyCoin = decreaseCount(copyCoin, unit);
+
+        if (inputCoin <= acc) {
+          setCoin(copyCoin);
+          return acc;
         }
+
+        const result = recursive(remainCoin - nearCoin);
+        return result;
       }
 
       return recursive(inputCoin);
     },
-    [coin, selectCoin]
+    [coin]
   );
 
   return { coin, selectCoin, correctCoin };
