+++ src/pages/Wallet/Wallet.styled.js
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const StyledWallet = styled.div`
+  width: 400px;
+  padding: 20px;
+  font-size: 1.2em;
+  font-weight: 700;
+  background-color: #ff8080;
+`;
+
+export { StyledWallet };
